[Gobblin-902][GOBBLIN-902] Enable gobblin yarn app luncher class configurable
Closes #2758 from ZihanLi58/GOBBLIN-902
